Move special.js and statistics.js to math folder.
